Update run.tcl with exit command
Add exit command to exit vitis at end of run.
Maybe it is because I am running this from a bash wrapper script, but vitis does not exit on completion of this script without this line. I am instancing vitis in shell mode with -f like on the lab webpage. Vitis 2022.2